refactored file repo. removed old query helpers.
