+++ src/components/authorization/utility/BaseNodeType.ts
@@ -1,26 +1,29 @@
 export enum BaseNodeType {
-  Project = 'Project',
-  ProjectMember = 'ProjectMember',
-  User = 'User',
   Budget = 'Budget',
   BudgetRecord = 'BudgetRecord',
-  Partnership = 'Partnership',
   Ceremony = 'Ceremony',
   Country = 'Country',
-  LanguageEngagement = 'LanguageEngagement',
-  InternshipEngagement = 'InternshipEngagement',
+  Directory = 'Directory',
+  Education = 'Education',
+  EthnologueLanguage = 'EthnologueLanguage',
+  File = 'File',
+  FileVersion = 'FileVersion',
   Film = 'Film',
   FundingAccount = 'FundingAccount',
   Language = 'Language',
-  EthnologueLanguage = 'EthnologueLanguage',
+  LanguageEngagement = 'LanguageEngagement',
   LiteracyMaterial = 'LiteracyMaterial',
-  Zone = 'Zone',
-  Region = 'Region',
+  InternshipEngagement = 'InternshipEngagement',
   Organization = 'Organization',
   Partner = 'Partner',
+  Partnership = 'Partnership',
   Product = 'Product',
+  Project = 'Project',
+  ProjectMember = 'ProjectMember',
+  Region = 'Region',
   Song = 'Song',
   Story = 'Story',
   Unavailability = 'Unavailability',
-  Education = 'Education',
+  User = 'User',
+  Zone = 'Zone',
 }
